quantum/requirements_quantum.txt
qiskit-ibm-runtime>=0.17.0
qiskit>=0.45.0
python-dotenv>=1.0.0
